+++ src/main/java/com/inq/learnease/dto/TaskReadResponseDto.java
@@ -1,6 +1,6 @@
 package com.inq.learnease.dto;
 
-import com.inq.learnease.entity.Task;
+import com.inq.learnease.entity.task.Task;
 
 import java.util.List;
