+++ src/main/java/com/inq/learnease/entity/task/Task.java
@@ -1,4 +1,4 @@
-package com.inq.learnease.entity;
+package com.inq.learnease.entity.task;
 
 import jakarta.persistence.*;
 import lombok.AccessLevel;
@@ -9,7 +9,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Entity
 @Getter
-@Table(name = "tasks")
+@Table(name = "task")
 public class Task {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
